Added checks for the ROOT_DIR argument.
Added struct ct_path and associated functions for encapsulating
the path to a file.  When a ct_path struct is created, its absolute
path is calculated and stat() is run to determine if the path
exists and if it is a directory.

When options are processed, check that the ROOT_DIR argument is
the path of an actual directory.