Improved ct_path_set functions.
The ct_path_set_add_path() function now checks that the path being added
is not equal to the root dir nor a duplicate of a path already in the set.
Paths in the path set are kept in sorted order.

Added ct_path_compare_path() function to aid in sorting of paths.

Added printed warnings if any PATH_TO_KEEP arguments are duplicates or
are equal to the root dir.

Added unit tests for various ct_path_set functions.

Added functional tests for permutations of three arguments.